Fixed logout issue
