+++ nursereports/client/pages/logout.py
@@ -1,13 +1,13 @@
 
 from .loading import loading_page
-from ...states.navbar import NavbarState
+from ...states.base import BaseState
 
 import reflex as rx
 
 @rx.page(
     route='/logout/[logout_reason]',
     title='Logging out...',
-    on_load=NavbarState.event_state_logout
+    on_load=BaseState.event_state_logout
 )
 def logout_page() -> rx.Component:
     return rx.box(
